requirements.txt
pyTelegramBotAPI
tensorflow
Pillow
numpy
python-dotenv
